Refactoring old Datamapper scripts, restoring HBS functionality, mending paths and imports
