+++ DSL/DataMapper/js/convert/base64ToText.js
@@ -1,4 +1,5 @@
-export default function base64ToText(base64String: string): string {
+export default function base64ToText(base64String) {
     const buffer = Buffer.from(base64String, 'base64');
     return buffer.toString('utf-8');
 }
+
