+++ DSL/DataMapper/js/file/fileWrite.js
@@ -1,20 +1,15 @@
-import express, { Router } from 'express';
-import fs from "fs";
-const router: Router = express.Router();
+import express from 'express';
+import fs from 'fs';
 
-interface FileRequestBody {
-    file_path: string;
-    content: string | Buffer;
-}
+const router = express.Router();
 
 router.post('/', (req, res) => {
-    const { file_path, content }: FileRequestBody = req.body;
+    const { file_path, content } = req.body;
 
     if (!file_path || !content) {
         res.status(400).send('Filename and content are required');
         return;
     }
-
 
     if (file_path.includes('..')) {
         res.status(400).send('Relative paths are not allowed');
@@ -30,4 +25,5 @@ router.post('/', (req, res) => {
         res.status(201).send('File saved successfully');
     });
 });
+
 export default router;
