+++ DSL/DataMapper/package.json
@@ -14,6 +14,9 @@
     "nodemailer": "^6.9.1",
     "jsdom": "^22.1.0",
     "replace-in-file": "^6.3.5",
-    "yaml": "^2.2.1"
+    "yaml": "^2.2.1",
+    "multer": "^1.4.5-lts.1",
+    "papaparse": "^5.3.2",
+    "lodash": "^4.14.191"
   }
 }
