--- DSL/DataMapper/ts/yamlToJson.ts
@@ -1,12 +1,0 @@
-import express, { Request, Response, Router } from 'express';
-import multer from "multer";
-import {parse} from "yaml";
-import base64ToText from "./base64ToText";
-const router: Router = express.Router();
-
-router.post('/', multer().array('file'), async (req: Request, res: Response) => {
-    const file = base64ToText(req.body.file);
-    let result = parse(file);
-    res.send(result);
-});
-export default router;
